完成CodeArgumentsPanel的改造
